requirements.txt
requests>=2.32.3
lxml>=4.9.3
schedule>=1.2.1
python-telegram-bot[job-queue]>=20.8
pyyaml>=6.0.1 